Start of refactoring to support multiple credential storages.
